+++ tests/test_doctests.py
@@ -8,7 +8,7 @@ PROJECT_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 
 @pytest.mark.parametrize("path", [
     os.path.join(path, filename)
-    for path, _, filenames in os.walk(PROJECT_ROOT)
+    for path, _, filenames in os.walk(os.path.join(PROJECT_ROOT, 'doc'))
     for filename in filenames
     if filename.endswith((".rst", ".md"))])
 def test_doctests(path):